% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_density.R
\name{kernel_density}
\alias{kernel_density}
\title{Kernel density estimation with plot functions}
\usage{
kernel_density(
  x,
  region,
  bdw = spatstat.core::bw.ppl,
  kernel = "gaussian",
  npixel = 100,
  plot = T,
  return_grid = F
)
}
\arguments{
\item{x}{`sf` point pattern.}

\item{region}{`sf` polygon that defines the study region.}

\item{bdw}{Bandwidth for kernel smoothing. Should be in projected values (meters
or feet). Can also provide any function that returns a bandwidth estimate. Default
behavior is to use `spatstat.core::bw.ppl`, but any of the base functions from the
spatstat family should work.}

\item{kernel}{Type of kernel to be used for the KDE estimation. Default is
'gaussian' but can also use 'epanechnikov', 'quartic', or 'disc'.}

\item{npixel}{Dimensions of grid points to be used.}

\item{plot}{Should a quick plot be generated?}

\item{return_grid}{Should the raster grid be returned? If TRUE, returns a
SpatialPixelsDataFrame, otherwise returns a dataframe with
columns for the density value, X, and Y. Defaults to FALSE.}
}
\description{
Computes a kernel density estimate for an `sf` point pattern.
}
\details{
The kernel_density function uses functionality in the `spatstat` package
to compute kernel density estimates for `sf` objects. The function can be used
to return a dataframe with density estimates for grid cells. This can further
be used for visualization purposes in `ggplot2`
}
\examples{

data("newhaven")
data("nh_hom")

# kernel density estimation with default bandwidth
kde_out <- kernel_density(x = nh_hom, region = newhaven)

head(kde_out)

# Plotting kde in ggplot
library(ggplot2)

ggplot(kde_out) +
 geom_tile(aes(x = X, y = Y, fill = density)) +
 coord_equal() +
 scale_fill_viridis_c() +
 theme_void()

}
